User CRUD without editing and validation.

+++ database/migrations/2020_08_06_065616_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
 
-			$table->integer('registration')->unsigned()->primary();
+			$table->bigInteger('registration')->unsigned()->primary();
 			$table->integer('category_id')->unsigned();
 			$table->string('name', 40);
 			$table->string('password', 255);
